act: separate class
